+++ libs/base/src/code/result/mod.rs
@@ -3,11 +3,13 @@
 #![doc = crate::_DOC_CODE_RESULT!()]
 //
 
+mod chain_hook; // Chain, Hook
 mod mismatch; // Mismatch
 mod reexports;
+mod value_quant; // ValueQuant
 
 crate::structural_mods! { // _mods
     _mods {
-        pub use super::{mismatch::*, reexports::*};
+        pub use super::{chain_hook::*, mismatch::*, reexports::*, value_quant::*};
     }
 }